Extract community events methods to module
Also use separate spec
