+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-04-22 20:01:06 +0300 using RuboCop version 0.67.2.
+# on 2019-04-23 10:58:55 +0300 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -15,13 +15,13 @@ Lint/UriEscapeUnescape:
 Metrics/AbcSize:
   Max: 23
 
-# Offense count: 14
+# Offense count: 15
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
   Max: 1248
 
-# Offense count: 5
+# Offense count: 4
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
   Max: 589
@@ -51,6 +51,7 @@ Naming/AccessorMethodName:
     - 'lib/teamlab/modules/community.rb'
     - 'lib/teamlab/modules/community/community_blogs.rb'
     - 'lib/teamlab/modules/community/community_bookmarks.rb'
+    - 'lib/teamlab/modules/community/community_events.rb'
     - 'lib/teamlab/modules/crm.rb'
     - 'lib/teamlab/modules/files.rb'
     - 'lib/teamlab/modules/group.rb'
@@ -107,7 +108,7 @@ Style/MutableConstant:
     - 'lib/teamlab/version.rb'
     - 'spec/support/http_data.rb'
 
-# Offense count: 287
+# Offense count: 288
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
