+++ spec/support/http_data.rb
@@ -1,4 +1,4 @@
-SERVER = 'https://nctautotest-rubygem-test.onlyoffice.com'.freeze
+SERVER = 'https://nctautotest-594591526-community-blogs-spec.teamlab.info'.freeze
 USERNAME = 'teamlab.ruby@gmail.com'.freeze
 PASSWORD = '123456'.freeze
